Added user authorization

+++ app/Http/Controllers/ListingController.php
@@ -65,6 +65,11 @@ class ListingController extends Controller
     // Update listing data
     public function update(Request $request, Listing $listing)
     {
+        // Make sure logged in user us owner
+        if ($listing->user_id != auth()->user()->id) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'company' => 'required',
             'title' => 'required',
@@ -87,6 +92,11 @@ class ListingController extends Controller
     // Delete listing
     public function destroy(Listing $listing)
     {
+        // Make sure logged in user us owner
+        if ($listing->user_id != auth()->user()->id) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully!');
